Implement book deletion and viewing of deleted books; enhance input handling and buffer management

+++ ValidIO.h
@@ -2,10 +2,11 @@
 #define VALIDIO_H
 
 void clearBuffer();
-int strInput(char*,int);
+int strInput(char*,size_t);
 int intInput(int*);
 int doubleInput(double*);
 void printErr();
 void comingSoon();
-
+// //void saveData(Book*,int*);
+// // void loadData(Book*,int*);
 #endif
